+++ lib/oli_web/controllers/lti_controller.ex
@@ -3,9 +3,15 @@ defmodule OliWeb.LtiController do
 
   alias Oli.Accounts
   alias Oli.Delivery.Sections
+  alias Oli.Institutions
+  alias Oli.Institutions.PendingRegistration
   alias Oli.Lti_1p3.ContextRoles
   alias Oli.Lti_1p3.PlatformRoles
   alias Oli.Lti_1p3
+  alias Oli.Predefined
+  alias Oli.Slack
+
+  require Logger
 
   ## LTI 1.3
   def login(conn, params) do
@@ -13,32 +19,32 @@ defmodule OliWeb.LtiController do
       {:ok, conn, redirect_url} ->
         conn
         |> redirect(external: redirect_url)
+      {:error, %{reason: :invalid_registration, msg: _msg, issuer: issuer, client_id: client_id}} ->
+        handle_invalid_registration(conn, issuer, client_id)
       {:error, reason} ->
         render(conn, "lti_error.html", reason: reason)
     end
   end
 
-  def launch(conn, _params) do
-    case Lti_1p3.LaunchValidation.validate(conn, &get_public_key/2) do
+  def launch(conn, _params \\ %{}) do
+    case Lti_1p3.LaunchValidation.validate(conn, providers: %{get_public_key: &get_public_key/2}) do
       {:ok, conn, lti_params} ->
-        deployment_id = lti_params["https://purl.imsglobal.org/spec/lti/claim/deployment_id"]
-        case Lti_1p3.get_ird_by_deployment_id(deployment_id) do
-          {institution, registration, deployment} ->
-            handle_valid_lti_1p3_launch(conn, lti_params, institution, registration, deployment)
-        end
-      {:error, %{reason: :invalid_deployment, msg: _msg, deployment_id: deployment_id}} ->
-        handle_no_deployment(conn, deployment_id)
+        handle_valid_lti_1p3_launch(conn, lti_params)
+      {:error, %{reason: :invalid_registration, msg: _msg, issuer: issuer, client_id: client_id}} ->
+        handle_invalid_registration(conn, issuer, client_id)
+      {:error, %{reason: :invalid_deployment, msg: _msg, registration_id: registration_id, deployment_id: deployment_id}} ->
+        handle_invalid_deployment(conn, registration_id, deployment_id)
       {:error, %{reason: _reason, msg: msg}} ->
-        render(conn, "basic_launch_invalid.html", reason: msg)
+        render(conn, "lti_error.html", reason: msg)
     end
   end
 
-  def test(conn, _params) do
-    case Lti_1p3.LaunchValidation.validate(conn, &get_public_key/2) do
+  def test(conn, _params \\ %{}) do
+    case Lti_1p3.LaunchValidation.validate(conn, providers: %{get_public_key: &get_public_key/2}) do
       {:ok, conn, lti_params} ->
         render(conn, "lti_test.html", lti_params: lti_params)
       {:error, %{reason: _reason, msg: msg}} ->
-        render(conn, "basic_launch_invalid.html", reason: msg)
+        render(conn, "lti_error.html", reason: msg)
     end
   end
 
@@ -58,13 +64,13 @@ defmodule OliWeb.LtiController do
 
     developer_key_config = %{
       "title" => "OLI Torus",
-      "scopes" => [],
-      ## TODO: add scopes for LTI services in the future
-      # "scopes":[
-      #   "https://purl.imsglobal.org/spec/lti-ags/scope/lineitem",
-      #   "https://purl.imsglobal.org/spec/lti-ags/scope/result.readonly"
-      #   ...
-      # ],
+      "scopes" => [
+        "https://purl.imsglobal.org/spec/lti-ags/scope/lineitem",
+        "https://purl.imsglobal.org/spec/lti-ags/scope/lineitem.readonly",
+        "https://purl.imsglobal.org/spec/lti-ags/scope/result.readonly",
+        "https://purl.imsglobal.org/spec/lti-ags/scope/score",
+        "https://purl.imsglobal.org/spec/lti-nrps/scope/contextmembership.readonly"
+      ],
       "extensions" => [
         %{
           "platform" => "canvas.instructure.com",
@@ -108,7 +114,7 @@ defmodule OliWeb.LtiController do
         "kty" => "RSA",
         "use" => "sig"
       },
-      "description" => "Create, deliver and iteratively improve course content with Torus, through the Open Learning Initiative",
+      "description" => "Create, deliver and iteratively improve course content through the Open Learning Initiative",
       "custom_fields" => %{},
       "public_jwk_url" => "https://#{host}/.well-known/jwks.json",
       "target_link_uri" => "https://#{host}/lti/launch",
@@ -142,82 +148,187 @@ defmodule OliWeb.LtiController do
 
   end
 
-  defp handle_no_deployment(conn, deployment_id) do
-    # TODO: add ability to configure a deployment from the LTI launch when the deployment does not exist
-    render(conn, "configure_deployment.html", deployment_id: deployment_id)
+  def request_registration(conn, %{"pending_registration" => pending_registration_attrs} = _params) do
+    case Institutions.create_pending_registration(pending_registration_attrs) do
+      {:ok, pending_registration} ->
+        # send a Slack notification regarding the new registration request
+        Slack.send(%{
+          "username" => "Torus Bot",
+          "icon_emoji" => ":robot_face:",
+          "blocks" => [
+            %{
+              "type" => "section",
+              "text" => %{
+                "type" => "mrkdwn",
+                "text" => "New registration request from *#{pending_registration.name}*. <#{Routes.institution_url(conn, :index)}#pending-registrations|Click here to view all pending requests>."
+              }
+            },
+            %{
+              "type" => "section",
+              "fields" => [
+                %{
+                  "type" => "mrkdwn",
+                  "text" => "*Name:*\n#{pending_registration.name}"
+                },
+                %{
+                  "type" => "mrkdwn",
+                  "text" => "*Institution Url:*\n#{pending_registration.institution_url}"
+                },
+                %{
+                  "type" => "mrkdwn",
+                  "text" => "*Contact Email:*\n#{pending_registration.institution_email}"
+                },
+                %{
+                  "type" => "mrkdwn",
+                  "text" => "*Location:*\n#{pending_registration.country_code} - #{pending_registration.timezone}"
+                },
+                %{
+                  "type" => "mrkdwn",
+                  "text" => "*Date:*\n#{pending_registration.inserted_at |> Timex.format!("{RFC822}")}"
+                },
+              ]
+            },
+            %{
+              "type" => "actions",
+              "elements" => [
+                %{
+                  "type" => "button",
+                  "text" => %{
+                    "type" => "plain_text",
+                    "text" => "Review Request"
+                  },
+                  "url" => "#{Routes.institution_url(conn, :index)}#pending-registrations",
+                }
+              ]
+            }
+          ]
+        })
+
+        conn
+        |> render("registration_pending.html")
+
+      {:error, changeset} ->
+        conn
+        |> render("register.html",
+          conn: conn,
+          changeset: changeset,
+          submit_action: Routes.lti_path(conn, :request_registration),
+          country_codes: Predefined.country_codes(),
+          timezones: Predefined.timezones(),
+          world_universities_and_domains: Predefined.world_universities_and_domains(),
+          lti_config_defaults: Predefined.lti_config_defaults(),
+          issuer: pending_registration_attrs["issuer"],
+          client_id: pending_registration_attrs["client_id"])
+    end
   end
 
-  defp handle_valid_lti_1p3_launch(conn, lti_params, institution, _registration, _deployment) do
-    lti_roles = lti_params["https://purl.imsglobal.org/spec/lti/claim/roles"]
+  defp handle_invalid_registration(conn, issuer, client_id) do
+    case Oli.Institutions.get_pending_registration_by_issuer_client_id(issuer, client_id) do
+      nil ->
+        conn
+        |> render("register.html",
+          conn: conn,
+          changeset: Institutions.change_pending_registration(%PendingRegistration{}),
+          submit_action: Routes.lti_path(conn, :request_registration),
+          country_codes: Predefined.country_codes(),
+          timezones: Predefined.timezones(),
+          world_universities_and_domains: Predefined.world_universities_and_domains(),
+          lti_config_defaults: Predefined.lti_config_defaults(),
+          issuer: issuer,
+          client_id: client_id)
 
-    # update user values defined by the oidc standard per LTI 1.3 standard user identity claims
-    # http://www.imsglobal.org/spec/lti/v1p3/#user-identity-claims
-    case Accounts.insert_or_update_user(%{
-      sub: lti_params["sub"],
-      name: lti_params["name"],
-      given_name: lti_params["given_name"],
-      family_name: lti_params["family_name"],
-      middle_name: lti_params["middle_name"],
-      nickname: lti_params["nickname"],
-      preferred_username:  lti_params["preferred_username"],
-      profile: lti_params["profile"],
-      picture: lti_params["picture"],
-      website: lti_params["website"],
-      email: lti_params["email"],
-      email_verified: lti_params["email_verified"],
-      gender: lti_params["gender"],
-      birthdate: lti_params["birthdate"],
-      zoneinfo: lti_params["zoneinfo"],
-      locale: lti_params["locale"],
-      phone_number: lti_params["phone_number"],
-      phone_number_verified: lti_params["phone_number_verified"],
-      address: lti_params["address"],
-      institution_id: institution.id,
-    }) do
-      {:ok, user} ->
-        # update user platform roles
-        Accounts.update_user_platform_roles(user, PlatformRoles.get_roles_by_uris(lti_roles))
+      _registration ->
+        conn
+        |> render("registration_pending.html")
+    end
+  end
 
-        # context claim is considered optional according to IMS http://www.imsglobal.org/spec/lti/v1p3/#context-claim
-        # safeguard against the case that context is missing
-        case lti_params["https://purl.imsglobal.org/spec/lti/claim/context"] do
-          nil ->
-            throw "Error getting context information from launch params"
-          context ->
-            %{"id" => context_id} = context
-            %{"title" => context_title} = context
+  defp handle_invalid_deployment(conn, registration_id, deployment_id) do
+    case Institutions.create_deployment(%{deployment_id: deployment_id, registration_id: registration_id}) do
+      {:ok, _deployment} ->
+        # try the LTI launch again now that deployment is created
+        launch(conn)
+      _ ->
+        render(conn, "lti_error.html", reason: "Failed to create deployment")
+    end
+  end
 
-            # Update section specifics - if one exists. Enroll the user and also update the section details
-            with {:ok, section} <- get_existing_section(context_id)
-            do
-              # transform lti_roles to a list only containing valid context roles (exclude all system and institution roles)
-              context_roles = ContextRoles.get_roles_by_uris(lti_roles)
+  defp handle_valid_lti_1p3_launch(conn, lti_params) do
+    deployment_id = lti_params["https://purl.imsglobal.org/spec/lti/claim/deployment_id"]
 
-              enroll_user(user.id, section.id, context_roles)
-              update_section_details(context_title, section)
+    case Lti_1p3.get_ird_by_deployment_id(deployment_id) do
+      {institution, _registration, _deployment} ->
+        lti_roles = lti_params["https://purl.imsglobal.org/spec/lti/claim/roles"]
+
+        # update user values defined by the oidc standard per LTI 1.3 standard user identity claims
+        # http://www.imsglobal.org/spec/lti/v1p3/#user-identity-claims
+        case Accounts.insert_or_update_user(%{
+          sub: lti_params["sub"],
+          name: lti_params["name"],
+          given_name: lti_params["given_name"],
+          family_name: lti_params["family_name"],
+          middle_name: lti_params["middle_name"],
+          nickname: lti_params["nickname"],
+          preferred_username:  lti_params["preferred_username"],
+          profile: lti_params["profile"],
+          picture: lti_params["picture"],
+          website: lti_params["website"],
+          email: lti_params["email"],
+          email_verified: lti_params["email_verified"],
+          gender: lti_params["gender"],
+          birthdate: lti_params["birthdate"],
+          zoneinfo: lti_params["zoneinfo"],
+          locale: lti_params["locale"],
+          phone_number: lti_params["phone_number"],
+          phone_number_verified: lti_params["phone_number_verified"],
+          address: lti_params["address"],
+          institution_id: institution.id,
+        }) do
+          {:ok, user} ->
+            # update user platform roles
+            Accounts.update_user_platform_roles(user, PlatformRoles.get_roles_by_uris(lti_roles))
+
+            # context claim is considered optional according to IMS http://www.imsglobal.org/spec/lti/v1p3/#context-claim
+            # safeguard against the case that context is missing
+            case lti_params["https://purl.imsglobal.org/spec/lti/claim/context"] do
+              nil ->
+                throw "Error getting context information from launch params"
+              context ->
+                %{"id" => context_id} = context
+                %{"title" => context_title} = context
+
+                # Update section specifics - if one exists. Enroll the user and also update the section details
+                with {:ok, section} <- get_existing_section(context_id)
+                do
+                  # transform lti_roles to a list only containing valid context roles (exclude all system and institution roles)
+                  context_roles = ContextRoles.get_roles_by_uris(lti_roles)
+
+                  enroll_user(user.id, section.id, context_roles)
+                  update_section_details(context_title, section)
+                end
+
+                # if account is linked to an author, sign them in
+                conn = if user.author_id != nil do
+                  author = Accounts.get_author!(user.author_id)
+
+                  # sign into authoring account using Pow
+                  conn
+                  |> use_pow_config(:author)
+                  |> Pow.Plug.create(author)
+                else
+                  conn
+                end
+
+                # sign current user in and redirect to home page
+                conn
+                |> use_pow_config(:user)
+                |> Pow.Plug.create(user)
+                |> redirect(to: Routes.delivery_path(conn, :index))
+
             end
-
-            # if account is linked to an author, sign them in
-            conn = if user.author_id != nil do
-              author = Accounts.get_author!(user.author_id)
-
-              # sign into authoring account using Pow
-              conn
-              |> use_pow_config(:author)
-              |> Pow.Plug.create(author)
-            else
-              conn
-            end
-
-            # sign current user in and redirect to home page
-            conn
-            |> use_pow_config(:user)
-            |> Pow.Plug.create(user)
-            |> redirect(to: Routes.delivery_path(conn, :index))
-
+            _ ->
+              throw "Error creating user"
         end
-        _ ->
-          throw "Error creating user"
     end
   end