+++ lib/oli_web/controllers/workspace_controller.ex
@@ -2,12 +2,10 @@ defmodule OliWeb.WorkspaceController do
   use OliWeb, :controller
   alias Oli.Authoring
   alias Oli.Accounts
-  alias Oli.Institutions
   alias Oli.Accounts.Author
 
   def account(conn, _params) do
     author = conn.assigns.current_author
-    institutions = Institutions.list_institutions() |> Enum.filter(fn i -> i.author_id == conn.assigns.current_author.id end)
     themes = Authoring.list_themes()
     active_theme = case author.preferences do
       nil ->
@@ -23,7 +21,6 @@ defmodule OliWeb.WorkspaceController do
       "account.html",
       title: "Account",
       active: :account,
-      institutions: institutions,
       themes: themes,
       preferences: author.preferences,
       active_theme: active_theme,