+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## Dev
+## Unreleased
 ### Enhancements
+  - Improved LTI workflow for new institutions
   - Add support for embedding images in structured content editors by external URL and by pasting a copied image
+  - Ordering activity
 
 ### Bug fixes
   - Fix a broken link to external learning objectives content